requirements.txt
aiohttp             3.8.1
aiosignal           1.2.0
async-timeout       4.0.2
attrs               21.4.0
base58              2.1.1
bitarray            2.5.1
certifi             2022.5.18.1
charset-normalizer  2.0.12
click               8.1.3
cytoolz             0.11.2
eth-abi             2.1.1
eth-account         0.5.8
eth-hash            0.3.2
eth-keyfile         0.5.1
eth-keys            0.3.4
eth-rlp             0.3.0
eth-typing          2.3.0
eth-utils           1.10.0
frozenlist          1.3.0
hexbytes            0.2.2
idna                3.3
importlib-resources 5.7.1
ipfshttpclient      0.8.0a2
jsonschema          4.6.0
lru-dict            1.1.7
multiaddr           0.0.9
multidict           6.0.2
netaddr             0.8.0
parsimonious        0.8.1
pip                 20.0.2
pkg-resources       0.0.0
protobuf            3.20.1
pycryptodome        3.14.1
pyrsistent          0.18.1
python-dotenv       0.20.0
requests            2.27.1
rlp                 2.0.1
setuptools          44.0.0
six                 1.16.0
toolz               0.11.2
uniswap-python      0.5.5
urllib3             1.26.9
varint              1.0.2
web3                5.29.2
websockets          9.1
yarl                1.7.2
zipp                3.8.0